# feat: optimise network history by removing staged segments as we go a…