closes #9547 

Before the flow was:
- fetch all segments
- load all segments into the database
- accidently forget to delete staged segments

Now it goes:
- fetch 1 segments
- load into database
- delete the staged segments
- move onto the next segment...

logs from locally auto-initialising network history from fairground:
```
INFO	datanode.networkHistory.service.store	store/store.go:679	staging full-segment	{"segment": "vega-fairground-202305051805-37-15278159-15279158.zip"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:355	copying into database	{"segment": "vega-fairground-202305051805-37-15278159-15279158.zip"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:361	copy complete	{"rows": 69534, "segment": "vega-fairground-202305051805-37-15278159-15279158.zip", "took": "1.151436311s"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:364	removing staged segment	{"path": "/Users/wwestgarth/basic_home/state/data-node/networkhistory/store/staging/vega-fairground-202305051805-37-15278159-15279158.zip"}
INFO	datanode.networkHistory.service.store	store/store.go:679	staging full-segment	{"segment": "vega-fairground-202305051805-37-15279159-15280158.zip"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:355	copying into database	{"segment": "vega-fairground-202305051805-37-15279159-15280158.zip"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:361	copy complete	{"rows": 71023, "segment": "vega-fairground-202305051805-37-15279159-15280158.zip", "took": "1.087890895s"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:364	removing staged segment	{"path": "/Users/wwestgarth/basic_home/state/data-node/networkhistory/store/staging/vega-fairground-202305051805-37-15279159-15280158.zip"}
INFO	datanode.networkHistory.service.store	store/store.go:679	staging full-segment	{"segment": "vega-fairground-202305051805-37-15280159-15281158.zip"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:355	copying into database	{"segment": "vega-fairground-202305051805-37-15280159-15281158.zip"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:361	copy complete	{"rows": 68972, "segment": "vega-fairground-202305051805-37-15280159-15281158.zip", "took": "1.105160871s"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:364	removing staged segment	{"path": "/Users/wwestgarth/basic_home/state/data-node/networkhistory/store/staging/vega-fairground-202305051805-37-15280159-15281158.zip"}
INFO	datanode.networkHistory.service.store	store/store.go:679	staging full-segment	{"segment": "vega-fairground-202305051805-37-15281159-15282158.zip"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:355	copying into database	{"segment": "vega-fairground-202305051805-37-15281159-15282158.zip"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:361	copy complete	{"rows": 68848, "segment": "vega-fairground-202305051805-37-15281159-15282158.zip", "took": "1.057671679s"}
INFO	datanode.networkHistory.service	snapshot/service_load_snapshot.go:364	removing staged segment	{"path": "/Users/wwestgarth/basic_home/state/data-node/networkhistory/store/staging/vega-fairground-202305051805-37-15281159-15282158.zip"}
```


system tests running the network history tests passing:
https://jenkins.vega.rocks/blue/organizations/jenkins/common%2Fsystem-tests-wrapper/detail/system-tests-wrapper/16221/pipeline/

except for one, which was a timing issue when the test wanted to assert a lot of things happened within the epoch. Running the test again and it passed:
https://jenkins.vega.rocks/blue/organizations/jenkins/common%2Fsystem-tests-wrapper/detail/system-tests-wrapper/16333/pipeline